modules,lib: fixed some header warnings discovered while testing mongodb

+++ lib/logmsg/logmsg.h
@@ -288,7 +288,7 @@ void log_msg_free_queue_node(LogMessageQueueNode *node);
 void log_msg_clear(LogMessage *self);
 void log_msg_merge_context(LogMessage *self, LogMessage **context, gsize context_len);
 
-LogMessage *log_msg_new(const gchar *msg, gint length,
+LogMessage *log_msg_new(const gchar *msg, gsize length,
                         GSockAddr *saddr,
                         MsgFormatOptions *parse_options);
 LogMessage *log_msg_new_mark(void);
